Start cleaning up analysis server code
